update vm_endpoints code
