+++ cli/cmd/vm_endpoints.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
+	"github.com/replicatedhq/replicated/pkg/platformclient"
 	"github.com/spf13/cobra"
 )
 
@@ -19,6 +20,8 @@ type VM struct {
 	DirectEndpoint string
 	DirectPort     int64
 	ID             string
+	Status         string
+	Name           string
 }
 
 // validateEndpointType validates that the endpoint type is supported
@@ -42,17 +45,24 @@ func (r *runners) InitVMSCPEndpoint(parent *cobra.Command) *cobra.Command {
 // initVMEndpointCmd creates a command for either SSH or SCP endpoint retrieval
 func (r *runners) initVMEndpointCmd(parent *cobra.Command, endpointType string) *cobra.Command {
 	protocol := strings.ToUpper(endpointType)
-	cmdUse := fmt.Sprintf("%s-endpoint VM_ID", endpointType)
+	cmdUse := fmt.Sprintf("%s-endpoint VM_ID_OR_NAME", endpointType)
 	cmdShort := fmt.Sprintf("Get the %s endpoint of a VM", protocol)
 
 	outputFormat := fmt.Sprintf("%s://username@hostname:port", endpointType)
 
 	cmdLong := fmt.Sprintf(`Get the %s endpoint and port of a VM.
 
-The output will be in the format: %s`, protocol, outputFormat)
+The output will be in the format: %s
+
+You can identify the VM either by its unique ID or by its name.
+
+Note: %s endpoints can only be retrieved from VMs in the "running" state.`, protocol, outputFormat, protocol)
 
 	cmdExample := fmt.Sprintf(`# Get %s endpoint for a specific VM by ID
-replicated vm %s-endpoint <id>`, protocol, endpointType)
+replicated vm %s-endpoint aaaaa11
+
+# Get %s endpoint for a specific VM by name
+replicated vm %s-endpoint my-test-vm`, protocol, endpointType, protocol, endpointType)
 
 	cmd := &cobra.Command{
 		Use:               cmdUse,
@@ -61,7 +71,7 @@ replicated vm %s-endpoint <id>`, protocol, endpointType)
 		Example:           cmdExample,
 		RunE:              r.VMEndpoint,
 		Args:              cobra.ExactArgs(1),
-		ValidArgsFunction: r.completeVMIDs,
+		ValidArgsFunction: r.completeVMIDsAndNames,
 		Annotations:       map[string]string{"endpointType": endpointType},
 	}
 
@@ -84,12 +94,16 @@ func (r *runners) VMEndpoint(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return r.getVMEndpoint(args[0], endpointType, nil, "")
+	vmID, err := r.getVMIDFromArg(args[0])
+	if err != nil {
+		return err
+	}
+
+	return r.getVMEndpoint(vmID, endpointType, nil, "")
 }
 
 // getVMEndpoint retrieves and formats VM endpoint with the specified protocol
 // endpointType should be either "ssh" or "scp"
-
 func (r *runners) getVMEndpoint(vmID, endpointType string, vm *VM, githubUsername string) error {
 	var err error
 
@@ -108,7 +122,13 @@ func (r *runners) getVMEndpoint(vmID, endpointType string, vm *VM, githubUsernam
 			DirectEndpoint: vmFromAPI.DirectSSHEndpoint,
 			DirectPort:     vmFromAPI.DirectSSHPort,
 			ID:             vmFromAPI.ID,
+			Status:         string(vmFromAPI.Status),
 		}
+	}
+
+	if vm.Status != "running" {
+		return errors.Errorf("VM %s is not in running state (current state: %s). %s is only available for running VMs",
+			vm.ID, vm.Status, strings.ToUpper(endpointType))
 	}
 
 	if vm.DirectEndpoint == "" || vm.DirectPort == 0 {
@@ -134,4 +154,42 @@ Visit %s to link your account`, githubAccountSettingsURL)
 	fmt.Printf("%s://%s@%s:%d\n", endpointType, githubUsername, vm.DirectEndpoint, vm.DirectPort)
 
 	return nil
+}
+
+// getVMIDFromArg resolves a VM ID or name to a VM ID
+func (r *runners) getVMIDFromArg(arg string) (string, error) {
+	_, err := r.kotsAPI.GetVM(arg)
+	if err == nil {
+		return arg, nil
+	}
+
+	cause := errors.Cause(err)
+	if cause != platformclient.ErrNotFound && cause != platformclient.ErrForbidden {
+		return "", errors.Wrap(err, "get vm")
+	}
+
+	vms, err := r.kotsAPI.ListVMs(false, nil, nil)
+	if errors.Cause(err) == platformclient.ErrForbidden {
+		return "", ErrCompatibilityMatrixTermsNotAccepted
+	} else if err != nil {
+		return "", errors.Wrap(err, "list vms")
+	}
+
+	var matchingVMs []string
+	for _, vm := range vms {
+		if vm.Name == arg {
+			matchingVMs = append(matchingVMs, vm.ID)
+		}
+	}
+
+	switch len(matchingVMs) {
+	case 0:
+		return "", errors.Errorf("VM with name or ID '%s' not found", arg)
+	case 1:
+		return matchingVMs[0], nil
+	default:
+		return "", errors.Errorf("Multiple VMs found with name '%s'. Please use the VM ID instead. Matching VMs: %s. To view all VM IDs run `replicated vm ls`",
+			arg,
+			fmt.Sprintf("%s (and %d more)", matchingVMs[0], len(matchingVMs)-1))
+	}
 }
