+++ cli/cmd/vm_endpoints_test.go
@@ -27,6 +27,7 @@ func TestGetVMEndpoint(t *testing.T) {
 				ID:             "vm-123",
 				DirectEndpoint: "test-vm.example.com",
 				DirectPort:     22,
+				Status:         "running",
 			},
 			mockGithubUsername: "testuser",
 			expectedOutput:     "ssh://testuser@test-vm.example.com:22\n",
@@ -40,6 +41,7 @@ func TestGetVMEndpoint(t *testing.T) {
 				ID:             "vm-456",
 				DirectEndpoint: "test-vm.example.com",
 				DirectPort:     22,
+				Status:         "running",
 			},
 			mockGithubUsername: "testuser",
 			expectedOutput:     "scp://testuser@test-vm.example.com:22\n",
@@ -53,6 +55,7 @@ func TestGetVMEndpoint(t *testing.T) {
 				ID:             "vm-789",
 				DirectEndpoint: "",
 				DirectPort:     0,
+				Status:         "running",
 			},
 			mockGithubUsername: "testuser",
 			expectedOutput:     "",
@@ -66,6 +69,7 @@ func TestGetVMEndpoint(t *testing.T) {
 				ID:             "vm-123",
 				DirectEndpoint: "test-vm.example.com",
 				DirectPort:     22,
+				Status:         "running",
 			},
 			mockGithubUsername: "",
 			expectedOutput:     "",
@@ -79,10 +83,25 @@ func TestGetVMEndpoint(t *testing.T) {
 				ID:             "vm-123",
 				DirectEndpoint: "test-vm.example.com",
 				DirectPort:     22,
+				Status:         "running",
 			},
 			mockGithubUsername: "testuser",
 			expectedOutput:     "",
 			expectedError:      "invalid endpoint type: invalid",
+		},
+		{
+			name:         "VM not in running state",
+			vmID:         "vm-123",
+			endpointType: "ssh",
+			mockVM: &VM{
+				ID:             "vm-123",
+				DirectEndpoint: "test-vm.example.com",
+				DirectPort:     22,
+				Status:         "stopped",
+			},
+			mockGithubUsername: "testuser",
+			expectedOutput:     "",
+			expectedError:      "VM vm-123 is not in running state (current state: stopped). SSH is only available for running VMs",
 		},
 	}
